103.118.42.104 ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEAvL7hONYANSdFyt/Qleaqcnp5LXuRn+YhDFDlnYxBJAVKs3ON8TQf9Psubq4ZO0IppKLs1jqh18ZyVLuKJ/94FblyuDI7RNtfqhy2pkpWdhQikgBdLwg3X4PYioEkJv6NEIybefCTwCyyqzJTjVFGPDTXxzaiSf7317FqFUabCN8xZZx6oGPI8FDszd65wQ7eP6BQQ9AIqDY/QFLRNhsEUxOwJasJ0LEzEZUXxoQoVvBckPvWQUXuqxVfRetCl40nkre0wa7PrzA/9y0rH1fkJV3NUGniYeOJUjkHBEBIZ9nR9M1SyDJC4Q7YBkQOMaghKv24VkPxnWmj8DfVkhpfTw== update

+++ stock/src/main/java/com/lic/stock/repository/BreakingPointRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lic.stock.domain.BreakingPointPO;
 
-public interface BreakingPointRepository extends JpaRepository<BreakingPointPO, Integer> {
+public interface BreakingPointRepository extends JpaRepository<BreakingPointPO, Long> {
     
 }
