+++ stock/src/main/java/com/lic/stock/repository/TradeDayRepository.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.lic.stock.domain.TradeDayPO;
 
-public interface TradeDayRepository extends JpaRepository<TradeDayPO, Integer> {
+public interface TradeDayRepository extends JpaRepository<TradeDayPO, Long> {
  
     public List<TradeDayPO> findBySymbolAndTradeDateBetween(String symbol,Date startDate, Date endDate);
